feat(heatmap): move stories to the dedicated storybook workspace
