+++ storybook/package.json
@@ -21,8 +21,10 @@
     "@nivo/calendar": "workspace:*",
     "@nivo/chord": "workspace:*",
     "@nivo/circle-packing": "workspace:*",
+    "@nivo/colors": "workspace:*",
     "@nivo/funnel": "workspace:*",
     "@nivo/generators": "workspace:*",
+    "@nivo/heatmap": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
